FEATURE: change model routine to include only intoxication level

+++ src/main/python/model/Utility.py
@@ -2,5 +2,5 @@ def determine_simulations(agent_parameter):
     """
     This function determines the number of iterations of the model.
     """
-    iterations = (agent_parameter["Maximum External Pressure"] - agent_parameter["Minimum External Pressure"]) / agent_parameter["External Pressure Increment"]
+    iterations = int((agent_parameter["Maximum External Pressure"] - agent_parameter["Minimum External Pressure"]) / agent_parameter["External Pressure Increment"])
     return iterations
